refactor Promotion-related classes; update applyVoucher method to use Long for subtotal, enhance discount calculation logic, and add PromotionUsage entity and repository

+++ src/main/java/com/manhduc205/ezgear/services/PromotionService.java
@@ -3,5 +3,5 @@ package com.manhduc205.ezgear.services;
 import java.math.BigDecimal;
 
 public interface PromotionService {
-    BigDecimal applyVoucher(String code, BigDecimal subtotal);
+    Long applyVoucher(String code, Long subtotal);
 }
